Add files via upload
This script deals with a nasty matter of fine-tuning BERT on a multinomial classification task. Here, the BERT uncased model is further trained based on the supplied data (in this case - political questions from different groups from the European Parliament). The performance of the model is not that great, but the code shows the overall workflow. Additional thanks goes to Rohan-Paul-Al (https://www.youtube.com/watch?v=4nNbg4bWDrQ) for his insightful video.